feat: migrate from Next.js to Vite
Replaces the Next.js Pages Router + static-export toolchain with a plain
React + Vite build. Runtime dependencies (Mantine, Monaco, reactflow, the
Globus SDK/components) are unchanged; only the build/routing/config layer
is swapped.

Build & tooling
- Remove `next` and `eslint-config-next`; add `vite`, `@vitejs/plugin-react`.
- Scripts: `dev` -> `vite`, `build` -> `tsc -b && vite build`, add `preview`,
  `lint` -> `eslint .` (flat config with typescript-eslint + react-hooks +
  react-refresh). Rule leniency mirrors the old `next/core-web-vitals`.
- Split tsconfig into the standard Vite `tsconfig.app.json` /
  `tsconfig.node.json` project references; `jsx` -> `react-jsx`.
- `next.config.mjs`, `next-env.d.ts`, `globals.d.ts`, `.eslintrc.json` removed.

Routing (multi-page build, no router dependency)
- Two HTML entries — `index.html` and `authenticate.html` — mirror the former
  static export. GitHub Pages serves `authenticate.html` at `/authenticate`,
  keeping the OAuth2 redirect URI working.
- `src/pages/_app.tsx` -> `src/Providers.tsx` (shared providers + one-time SDK
  setup); `pages/index.tsx` -> `App.tsx`; `pages/authenticate.tsx` ->
  `AuthenticateRoute.tsx`; `pages/_document.tsx` dropped (markup now in HTML).
- `FlowDefinition` moved to `src/flow.ts` (was exported from `pages/index`).
- `next/router` / `next/navigation` usages replaced with `history.replaceState`
  / `location.assign` / `location.replace`. `next/head` metadata moved into the
  HTML entries.

Config / env
- Env vars renamed `NEXT_PUBLIC_*` -> `VITE_*`; base path is now Vite's `base`
  (from `VITE_BASE_PATH`), and code uses `import.meta.env.BASE_URL` for assets
  and the redirect URI.
- SST `StaticSite` output `out` -> `dist`, env `VITE_BASE_PATH: "/"`.
- GitHub Pages workflow `nextjs.yml` -> `pages.yml` (Vite build, uploads
  `dist`); `release-please.yml` and `dependabot.yml` updated accordingly.

Verified: `npm run build`, `npm run lint`, and the dev server (both routes
serve 200 under the `/flows-ide` base path) all pass.
