+++ .env
@@ -1,5 +1,5 @@
-NEXT_PUBLIC_BASE_PATH=/flows-ide
+VITE_BASE_PATH=/flows-ide
 
 
-NEXT_PUBLIC_GLOBUS_CLIENT_ID="91847941-14e5-4dcd-acf9-63f53741def8"
-NEXT_PUBLIC_GLOBUS_SCOPES="https://auth.globus.org/scopes/eec9b274-0c81-4334-bdc2-54e90e689b9a/manage_flows"
+VITE_GLOBUS_CLIENT_ID="91847941-14e5-4dcd-acf9-63f53741def8"
+VITE_GLOBUS_SCOPES="https://auth.globus.org/scopes/eec9b274-0c81-4334-bdc2-54e90e689b9a/manage_flows"
