+++ .env.development
@@ -1,3 +1,3 @@
-NEXT_PUBLIC_GLOBUS_ENVIRONMENT="sandbox"
+VITE_GLOBUS_ENVIRONMENT="sandbox"
 # https://app.sandbox.globuscs.info/settings/developers/projects/9a0b94e6-07be-48d1-911e-6cdab241d18d/apps/0cc46409-940b-4bd4-b920-d6aceecf40cf
-NEXT_PUBLIC_GLOBUS_CLIENT_ID="0cc46409-940b-4bd4-b920-d6aceecf40cf"
+VITE_GLOBUS_CLIENT_ID="0cc46409-940b-4bd4-b920-d6aceecf40cf"
