--- next.config.mjs
@@ -1,11 +1,0 @@
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-  reactStrictMode: true,
-  output: "export",
-  basePath: process.env.NEXT_PUBLIC_BASE_PATH,
-  images: {
-    unoptimized: true,
-  },
-};
-
-export default nextConfig;
