+++ package.json
@@ -2,11 +2,12 @@
   "name": "@globus/flows-ide",
   "version": "2.1.1",
   "private": true,
+  "type": "module",
   "scripts": {
-    "build": "next build",
-    "dev": "next dev --turbopack",
-    "lint": "next lint",
-    "start": "next start"
+    "build": "tsc -b && vite build",
+    "dev": "vite",
+    "preview": "vite preview",
+    "lint": "eslint ."
   },
   "overrides": {
     "@globus/sdk": "6.2.0-canary.85.ba009d1"
@@ -23,7 +24,6 @@
     "@monaco-editor/react": "^4.7.0",
     "lodash": "^4.18.1",
     "lz-string": "^1.5.0",
-    "next": "^15.5.7",
     "react": "19.2.5",
     "react-dom": "19.2.5",
     "react-icons": "^5.6.0",
@@ -31,19 +31,25 @@
     "zustand": "^5.0.12"
   },
   "devDependencies": {
+    "@eslint/js": "^9.0.0",
     "@types/lodash": "^4.17.24",
     "@types/node": "^24",
     "@types/react": "19.2.14",
     "@types/react-dom": "19.2.3",
-    "eslint": "^8",
-    "eslint-config-next": "15.5.2",
+    "@vitejs/plugin-react": "^5.0.0",
+    "eslint": "^9.0.0",
     "eslint-config-prettier": "^10.1.8",
+    "eslint-plugin-react-hooks": "^5.0.0",
+    "eslint-plugin-react-refresh": "^0.4.0",
+    "globals": "^16.0.0",
     "postcss": "^8.5.10",
     "postcss-preset-mantine": "^1.18.0",
     "postcss-simple-vars": "^7.0.1",
     "prettier": "^3.8.3",
     "sst": "^4.12.2",
-    "typescript": "~6.0.3"
+    "typescript": "~6.0.3",
+    "typescript-eslint": "^8.0.0",
+    "vite": "^7.0.0"
   },
   "packageManager": "npm@11.5.1"
 }
