+++ src/components/DocumentationBrowser/DocumentationBrowser.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 import {
   Anchor,
   Accordion,
