+++ src/flow.ts
@@ -1,0 +1,19 @@
+export type FlowDefinition = {
+  States: {
+    [key: string]: {
+      Type: "Action" | "Choice" | "Fail" | "Pass" | "ExpressionEval" | string;
+      Next?: string;
+      End?: boolean;
+      Comment?: string;
+      Catch?: {
+        Next?: string;
+      }[];
+      Choices?: {
+        Next?: string;
+      }[];
+      [key: string]: any;
+    };
+  };
+  StartAt: string;
+  Comment?: string;
+};
